:art: Improving structure of the package
