+++ kiss_headers/__init__.py
@@ -35,7 +35,8 @@ at <https://github.com/Ousret/kiss-headers>.
 :license: MIT, see LICENSE for more details.
 """
 
-from kiss_headers.utils import Headers, Header, parse_it, lock_output_type
+from kiss_headers.models import Headers, Header
+from kiss_headers.api import parse_it, lock_output_type
 from kiss_headers.builder import (
     CustomHeader,
     ContentType,
